cleanup application structure
