+++ app/components/SchoolInfo.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../actions/index'
-import styles from './styles/home.css'
+import styles from './styles/schoolinfo.css'
 
 const _ = require('lodash')
