+++ app/components/exam/ExamListInputs.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/list.css'
+import styles from './styles/exam.css'
 
 const _ = require('lodash')
