+++ app/components/exam/Exams.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/exam.css'
+import styles from './styles/exam.css'
 import ExamForm from './ExamForm'
 import ExamList from './ExamList'
