+++ app/components/exam/styles/exam.css
@@ -49,3 +49,19 @@
   width: 170px;
   display: inline-block;
 }
+
+.list_btn {
+  display: inline-flex;
+  align-items: flex-start;
+  height: 40px;
+}
+
+.badge_number {
+  float: right;
+}
+
+.exam_div {
+  margin: 10px 10px 20px 10px;
+  height: 300px;
+  overflow: scroll;
+}
