+++ app/components/grades/GradeTable.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import ReactTable from 'react-table'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/grades.css'
+import styles from './styles/grades.css'
 import { gradeColumns } from './helpers/gradeColumns'
 import { gradeInfo } from './helpers/gradeInfo'
 import GradeDropdown from './GradeDropdown'
