+++ app/components/rooms/ClassList.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/list.css'
+import styles from './styles/room.css'
 import RoomModal from './RoomModal'
 
 const _ = require('lodash')
