+++ app/components/rooms/Navbar.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/classroom.css'
+import styles from './styles/room.css'
 
 const NavBar = ({ tabStatus, actions }) => (
 	<div className={styles.navbar_div}>
