+++ app/components/rooms/RoomModal.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { Modal, ModalHeader, ModalBody, ModalFooter, Button } from 'reactstrap'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/room.css'
+import styles from './styles/room.css'
 
 const _ = require('lodash')
