+++ app/components/students/StudentForm.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
-import styles from '../styles/students.css'
+import styles from './styles/students.css'
 import studentForm from './helpers/formHelper'
 
 const _ = require('lodash')
