+++ app/components/students/styles/students.css
@@ -18,7 +18,7 @@
 .form_inner_div {
   white-space: nowrap;
   float: right;
-  margin: 10px 0 0 0;
+  margin: 10px -15px 0 0;
   width: 300px;
 }
