+++ app/components/styles/SideMenu.css
@@ -16,4 +16,3 @@
   border: none;
   background: #007bff;
 }
-
