--- app/components/styles/classroom.css
@@ -1,51 +1,0 @@
-.navbar_div {
-  padding: 30px 0 5px 0;
-}
-
-.room_div {
-  white-space: 800px;
-  border-top: 1px solid;
-  padding: 0 20px 0 0;
-}
-
-.form_div {
-  white-space: nowrap;
-  float: right;
-  margin: 10px 0 0 0;
-}
-
-input[type='text'] {
-  border: 0;
-  outline: 0;
-  float: left;
-  background: transparent;
-  border-bottom: 1px solid white;
-  width: 200px;
-}
-
-.save_btn {
-  float: right;
-  margin: 10px 0 0 200px;
-}
-
-.subject_save {
-  float: right;
-  margin: 10px 0 0 230px;
-}
-
-.form_outer_div {
-  float: right;
-  height: 350px;
-  width: 500px;
-}
-
-.form_label {
-  float: left;
-  margin-right: 5px;
-  vertical-align: bottom;
-  line-height: 32px;
-}
-
-.center_header {
-  text-align: center;
-}
