--- app/components/styles/list.css
@@ -1,20 +1,0 @@
-.list_div {
-  height: 230px;
-  overflow: scroll;
-}
-
-.list_btn {
-  display: inline-flex;
-  align-items: flex-start;
-  height: 40px;
-}
-
-.badge_number {
-  float: right;
-}
-
-.exam_div {
-  margin: 10px 10px 20px 10px;
-  height: 300px;
-  overflow: scroll;
-}
