--- app/components/styles/room.css
@@ -1,15 +1,0 @@
-.form_label {
-  float: left;
-  margin: 3px 10px 0 0;
-  vertical-align: bottom;
-  line-height: 32px;
-}
-
-.form_input {
-  float: right;
-}
-
-.form_div {
-  white-space: nowrap;
-  display: inline-block;
-}
