+++ app/components/styles/schoolinfo.css
@@ -16,6 +16,7 @@
 input[type='text'] {
   border: 0;
   outline: 0;
+  background: transparent;
   border-bottom: 1px solid white;
   width: 200px;
 }
