+++ app/components/subject/styles/subject.css
@@ -1,0 +1,70 @@
+.form_div {
+  white-space: nowrap;
+  float: right;
+  margin: 10px 0 0 0;
+}
+
+.form_label {
+  float: left;
+  margin-right: 5px;
+  vertical-align: bottom;
+  line-height: 32px;
+}
+
+.subject_save {
+  float: right;
+  margin: 10px 0 0 230px;
+}
+
+.list_div {
+  height: 230px;
+  overflow: scroll;
+}
+
+.list_btn {
+  display: inline-flex;
+  align-items: flex-start;
+  height: 40px;
+}
+
+.badge_number {
+  float: right;
+}
+
+.modal_form_div {
+  white-space: nowrap;
+  display: inline-block;
+}
+
+.modal_form_label {
+  float: left;
+  margin: 3px 10px 0 0;
+  vertical-align: bottom;
+  line-height: 32px;
+}
+
+.subject_left {
+  float: left;
+  width: 38%;
+  height: 330px;
+  position: absolute;
+  overflow: scroll;
+  margin-left: 110px;
+}
+
+.subject_right {
+  width: 40%;
+  height: 330px;
+  position: absolute;
+  right: 20px;
+  overflow: scroll;
+}
+
+.center_header {
+  text-align: center;
+}
+
+.main_div {
+  margin-top: 30px;
+  border-top: 1px solid;
+}
