+++ app/containers/Home.js
@@ -3,28 +3,28 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../actions/index'
 import SideMenu from '../components/SideMenu'
-import SchoolPage from '../components/SchoolPage'
+import Homepage from '../components/Homepage'
 
 class Home extends Component {
-  componentDidMount() {
-    this.props.actions.displaySchoolData()
-  }
+	componentDidMount() {
+		this.props.actions.displaySchoolData()
+	}
 
-  render() {
-    return (
-      <div data-tid="container">
-        <SideMenu />
-        <SchoolPage />
-      </div>
-    )
-  }
+	render() {
+		return (
+			<div data-tid="container">
+				<SideMenu />
+				<Homepage />
+			</div>
+		)
+	}
 }
 
 const mapDispatchToProps = dispatch => ({
-  actions: bindActionCreators(actionCreators, dispatch)
+	actions: bindActionCreators(actionCreators, dispatch)
 })
 
 export default connect(
-  null,
-  mapDispatchToProps
+	null,
+	mapDispatchToProps
 )(Home)
