+++ app/containers/Root.js
@@ -1,23 +1,23 @@
-import React, { Component } from 'react';
-import { Provider } from 'react-redux';
-import { ConnectedRouter } from 'connected-react-router';
-import type { Store } from '../reducers/types';
-import Routes from '../Routes';
+import React, { Component } from 'react'
+import { Provider } from 'react-redux'
+import { ConnectedRouter } from 'connected-react-router'
+import type { Store } from '../reducers/types'
+import Routes from '../Routes'
 
 type Props = {
-  store: Store,
-  history: {}
-};
+	store: Store,
+	history: {}
+}
 
 export default class Root extends Component<Props> {
-  render() {
-    const { store, history } = this.props;
-    return (
-      <Provider store={store}>
-        <ConnectedRouter history={history}>
-          <Routes />
-        </ConnectedRouter>
-      </Provider>
-    );
-  }
+	render() {
+		const { store, history } = this.props
+		return (
+			<Provider store={store}>
+				<ConnectedRouter history={history}>
+					<Routes />
+				</ConnectedRouter>
+			</Provider>
+		)
+	}
 }
